+++ cms/globals/About.ts
@@ -1,0 +1,83 @@
+import type { GlobalConfig } from 'payload';
+
+export const About: GlobalConfig = {
+  slug: 'about',
+  access: {
+    read: () => true,
+  },
+  fields: [
+    {
+      name: 'socialLinks',
+      type: 'array',
+      fields: [
+        {
+          name: 'platform',
+          type: 'text',
+          required: true,
+        },
+        {
+          name: 'url',
+          type: 'text',
+          required: true,
+        },
+        {
+          name: 'icon',
+          type: 'text',
+          required: true,
+        },
+        {
+          name: 'showLabel',
+          type: 'checkbox',
+          required: true,
+          defaultValue: false,
+        }
+      ],
+    },
+    {
+      name: 'experiences',
+      type: 'array',
+      fields: [
+        {
+          name: 'company',
+          type: 'text',
+          required: true,
+        },
+        {
+          name: 'position',
+          type: 'text',
+          required: true,
+        },
+        {
+          name: 'startDate',
+          type: 'date',
+          required: true,
+        },
+        {
+          name: 'endDate',
+          type: 'date',
+        },
+        {
+          name: 'responsibilities',
+          type: 'richText',
+          required: true,
+        },
+      ],
+    },
+    {
+      name: 'skills',
+      type: 'array',
+      fields: [
+        {
+          name: 'name',
+          type: 'text',
+          required: true,
+        },
+        {
+          name: 'icon',
+          type: 'text',
+          required: true,
+        },
+      ],
+    },
+  ],
+};