+++ src/components/Icon.vue
@@ -1,10 +1,13 @@
 <template>
   <svg :width="width" :height="height" class="icon">
-    <use :href="`/icons/sprite.svg#${name}`" />
+    <use :href="`/icons/sprite.svg#${iconName}`" />
   </svg>
 </template>
 
 <script setup>
+import { computed } from 'vue';
+import { useTheme } from '@/composables/useTheme';
+
 const props = defineProps({
   name: String,
   width: {
@@ -15,6 +18,12 @@ const props = defineProps({
     type: [String, Number],
     default: 24
   }
+});
+
+const { currentTheme } = useTheme();
+
+const iconName = computed(() => {
+  return currentTheme.value === 'dark' ? `${props.name}-light` : props.name;
 });
 </script>